@@ -1,28 +1,26 @@
 import { useContext, useEffect, useState } from 'react';
-import { BrowserRouter as Router, Switch, Redirect } from 'react-router-dom';
+import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import { getAuth, onAuthStateChanged } from '@firebase/auth';
 
 import { PrivateRoute } from './PrivateRoute';
 import { PublicRoute } from './PublicRoute';
-import { AuthRouter } from './AuthRouter';
+import { AuthRouter } from './AuthRouter.js';
 import PinterestScreen from '../components/pinterestScreen/PinterestScreen';
 import AuthContext from '../context/authContext/AuthContext';
+import NotFound from '../components/pinterestScreen/NotFound';
 
 export const AppRouter = () => {
-	const { login } = useContext(AuthContext);
+	const { login, startLogOut } = useContext(AuthContext);
 	const [checking, setChecking] = useState(true);
 
-	const [isLoggedIn, setIsLoggedIn] = useState(false);
-
-	//Efecto para regresar el usuario autenticado
+	// Efecto para regresar el usuario autenticado
 	useEffect(() => {
 		const auth = getAuth();
 		onAuthStateChanged(auth, (user) => {
 			if (user?.uid) {
 				login(user.uid, user.displayName);
-				setIsLoggedIn(true);
 			} else {
-				setIsLoggedIn(false);
+				startLogOut();
 			}
 			setChecking(false);
 		});
@@ -34,25 +32,26 @@ export const AppRouter = () => {
 	}
 
 	return (
-		<Router>
-			<div>
-				<Switch>
-					<PublicRoute
-						isAuthenticated={isLoggedIn}
-						path='/auth'
-						component={AuthRouter}
-					/>
-
-					<PrivateRoute
-						isAuthenticated={isLoggedIn}
-						exact
-						path='/'
-						component={PinterestScreen}
-					/>
-
-					<Redirect to='/auth/login' />
-				</Switch>
-			</div>
-		</Router>
+		<BrowserRouter>
+			<Routes>
+				<Route
+					path='/auth/*'
+					element={
+						<PublicRoute>
+							<AuthRouter />
+						</PublicRoute>
+					}
+				/>
+				<Route
+					path='/'
+					element={
+						<PrivateRoute>
+							<PinterestScreen />
+						</PrivateRoute>
+					}
+				/>
+				<Route path='/*' element={<NotFound />} />
+			</Routes>
+		</BrowserRouter>
 	);
 };
